refactor : interface 변경 및 서비스 이름 변경
